order_reformer_function/requirements.txt
apimatic-core==0.2.6
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
attrs==22.2.0
boto3==1.26.109
botocore==1.29.109
CacheControl==0.12.14
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==40.0.1
deprecation==2.1.0
enum34==1.1.10
exceptiongroup==1.1.0
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
jsonpath-ng==1.6.0
jsonpickle==3.0.2
jsonpointer==2.4
msgpack==1.0.7
packaging==22.0
pluggy==1.0.0
ply==3.11
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.0
six==1.16.0
squareup==33.1.0.20231213
tomli==2.0.1
urllib3==1.26.13
